ChoiceRule class

Factory class for creating a choice rule.

Methods

Public methods

ChoiceRule$StringEquals()

ChoiceRule$StringEqualsPath()

ChoiceRule$StringLessThan()

ChoiceRule$StringLessThanPath()

ChoiceRule$StringGreaterThan()

ChoiceRule$StringGreaterThanPath()

ChoiceRule$StringLessThanEquals()

ChoiceRule$StringLessThanEqualsPath()

ChoiceRule$StringGreaterThanEquals()

ChoiceRule$StringGreaterThanEqualsPath()

ChoiceRule$NumericEquals()

ChoiceRule$NumericEqualsPath()

ChoiceRule$NumericLessThan()

ChoiceRule$NumericLessThanPath()

ChoiceRule$NumericGreaterThan()

ChoiceRule$NumericGreaterThanPath()

ChoiceRule$NumericLessThanEquals()

ChoiceRule$NumericLessThanEqualsPath()

ChoiceRule$NumericGreaterThanEquals()

ChoiceRule$NumericGreaterThanEqualsPath()

ChoiceRule$BooleanEquals()

ChoiceRule$BooleanEqualsPath()

ChoiceRule$TimestampEquals()

ChoiceRule$TimestampEqualsPath()

ChoiceRule$TimestampLessThan()

ChoiceRule$TimestampLessThanPath()

ChoiceRule$TimestampGreaterThan()

ChoiceRule$TimestampGreaterThanPath()

ChoiceRule$TimestampLessThanEquals()

ChoiceRule$TimestampLessThanEqualsPath()

ChoiceRule$TimestampGreaterThanEquals()

ChoiceRule$TimestampGreaterThanEqualsPath()

ChoiceRule$IsNull()

ChoiceRule$IsPresent()

ChoiceRule$IsString()

ChoiceRule$IsNumeric()

ChoiceRule$IsTimestamp()

ChoiceRule$IsBoolean()

ChoiceRule$StringMatches()

ChoiceRule$And()

ChoiceRule$Or()

ChoiceRule$Not()

ChoiceRule$clone()

Method StringEquals()

Creates a rule with the `StringEquals` operator.

Usage

ChoiceRule$StringEquals(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(str): Constant value to compare `variable` against.

Returns

Rule: Rule with `StringEquals` operator.

Method StringEqualsPath()

Creates a rule with the `StringEqualsPath` operator.

Usage

ChoiceRule$StringEqualsPath(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(str): Path to the value to compare `variable` against.

Returns

Rule: Rule with `StringEqualsPath` operator.

Method StringLessThan()

Creates a rule with the `StringLessThan` operator.

Usage

ChoiceRule$StringLessThan(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(str): Constant value to compare `variable` against.

Returns

Rule: Rule with `StringLessThan` operator.

Method StringLessThanPath()

Creates a rule with the `StringLessThanPath` operator.

Usage

ChoiceRule$StringLessThanPath(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(str): Path to the value to compare `variable` against.

Returns

Rule: Rule with `StringLessThanPath` operator.

Method StringGreaterThan()

Creates a rule with the `StringGreaterThan` operator.

Usage

ChoiceRule$StringGreaterThan(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(str): Constant value to compare `variable` against.

Returns

Rule: Rule with `StringGreaterThan` operator.

Method StringGreaterThanPath()

Creates a rule with the `StringGreaterThanPath` operator.

Usage

ChoiceRule$StringGreaterThanPath(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(str): Path to the value to compare `variable` against.

Returns

Rule: Rule with `StringGreaterThanPath` operator

Method StringLessThanEquals()

Creates a rule with the `StringLessThanEquals` operator.

Usage

ChoiceRule$StringLessThanEquals(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(str): Constant value to compare `variable` against.

Returns

Rule: Rule with `StringLessThanEquals` operator.

Method StringLessThanEqualsPath()

Creates a rule with the `StringLessThanEqualsPath` operator.

Usage

ChoiceRule$StringLessThanEqualsPath(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(str): Path to the value to compare `variable` against.

Returns

Rule: Rule with `StringLessThanEqualsPath` operator.

Method StringGreaterThanEquals()

Creates a rule with the `StringGreaterThanEquals` operator.

Usage

ChoiceRule$StringGreaterThanEquals(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(str): Constant value to compare `variable` against.

Returns

Rule: Rule with `StringGreaterThanEquals` operator.

Method StringGreaterThanEqualsPath()

Creates a rule with the `StringGreaterThanEqualsPath` operator.

Usage

ChoiceRule$StringGreaterThanEqualsPath(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(str): Path to the value to compare `variable` against.

Returns

Rule: Rule with `StringGreaterThanEqualsPath` operator.

Method NumericEquals()

Creates a rule with the `NumericEquals` operator.

Usage

ChoiceRule$NumericEquals(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(int): Constant value to compare `variable` against.

Returns

Rule: Rule with `NumericEquals` operator.

Method NumericEqualsPath()

Creates a rule with the `NumericEqualsPath` operator.

Usage

ChoiceRule$NumericEqualsPath(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(str): Path to the value to compare `variable` against.

Returns

Rule: Rule with `NumericEqualsPath` operator.

Method NumericLessThan()

Creates a rule with the `NumericLessThan` operator.

Usage

ChoiceRule$NumericLessThan(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(int): Constant value to compare `variable` against.

Returns

Rule: Rule with `NumericLessThan` operator.

Method NumericLessThanPath()

Creates a rule with the `NumericLessThanPath` operator.

Usage

ChoiceRule$NumericLessThanPath(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(str): Path to the value to compare `variable` against.

Returns

Rule: Rule with `NumericLessThanPath` operator.

Method NumericGreaterThan()

Creates a rule with the `NumericGreaterThan` operator.

Usage

ChoiceRule$NumericGreaterThan(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(int): Constant value to compare `variable` against.

Returns

Rule: Rule with `NumericGreaterThan` operator.

Method NumericGreaterThanPath()

Creates a rule with the `NumericGreaterThanPath` operator.

Usage

ChoiceRule$NumericGreaterThanPath(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(str): Path to the value to compare `variable` against.

Returns

Rule: Rule with `NumericGreaterThanPath` operator.

Method NumericLessThanEquals()

Creates a rule with the `NumericLessThanEquals` operator.

Usage

ChoiceRule$NumericLessThanEquals(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(int): Constant value to compare `variable` against.

Returns

Rule: Rule with `NumericLessThanEquals` operator.

Method NumericLessThanEqualsPath()

Creates a rule with the `NumericLessThanEqualsPath` operator.

Usage

ChoiceRule$NumericLessThanEqualsPath(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(str): Path to the value to compare `variable` against.

Returns

Rule: Rule with `NumericLessThanEqualsPath` operator

Method NumericGreaterThanEquals()

Creates a rule with the `NumericGreaterThanEquals` operator.

Usage

ChoiceRule$NumericGreaterThanEquals(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(int): Constant value to compare `variable` against.

Returns

Rule: Rule with `NumericGreaterThanEquals` operator.

Method NumericGreaterThanEqualsPath()

Creates a rule with the `NumericGreaterThanEqualsPath` operator.

Usage

ChoiceRule$NumericGreaterThanEqualsPath(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(str): Path to the value to compare `variable` against.

Returns

Rule: Rule with `NumericGreaterThanEqualsPath` operator.

Method BooleanEquals()

Creates a rule with the `BooleanEquals` operator.

Usage

ChoiceRule$BooleanEquals(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(bool): Constant value to compare `variable` against.

Returns

Rule: Rule with `BooleanEquals` operator.

Method BooleanEqualsPath()

Creates a rule with the `BooleanEqualsPath` operator.

Usage

ChoiceRule$BooleanEqualsPath(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(str): Path to the value to compare `variable` against.

Returns

Rule: Rule with `BooleanEqualsPath` operator.

Method TimestampEquals()

Creates a rule with the `TimestampEquals` operator.

Usage

ChoiceRule$TimestampEquals(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(str): Constant value to compare `variable` against.

Returns

Rule: Rule with `TimestampEquals` operator.

Method TimestampEqualsPath()

Creates a rule with the `TimestampEqualsPath` operator.

Usage

ChoiceRule$TimestampEqualsPath(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(str): Path to the value to compare `variable` against.

Returns

Rule: Rule with `TimestampEqualsPath` operator.

Method TimestampLessThan()

Creates a rule with the `TimestampLessThan` operator.

Usage

ChoiceRule$TimestampLessThan(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(str): Constant value to compare `variable` against.

Returns

Rule: Rule with `TimestampLessThan` operator.

Method TimestampLessThanPath()

Creates a rule with the `TimestampLessThanPath` operator.

Usage

ChoiceRule$TimestampLessThanPath(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(str): Path to the value to compare `variable` against.

Returns

Rule: Rule with `TimestampLessThanPath` operator.

Method TimestampGreaterThan()

Creates a rule with the `TimestampGreaterThan` operator.

Usage

ChoiceRule$TimestampGreaterThan(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(str): Constant value to compare `variable` against.

Returns

Rule: Rule with `TimestampGreaterThan` operator.

Method TimestampGreaterThanPath()

Creates a rule with the `TimestampGreaterThanPath` operator.

Usage

ChoiceRule$TimestampGreaterThanPath(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(str): Path to the value to compare `variable` against.

Returns

Rule: Rule with `TimestampGreaterThanPath` operator.

Method TimestampLessThanEquals()

Creates a rule with the `TimestampLessThanEquals` operator.

Usage

ChoiceRule$TimestampLessThanEquals(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(str): Constant value to compare `variable` against.

Returns

Rule: Rule with `TimestampLessThanEquals` operator.

Method TimestampLessThanEqualsPath()

Creates a rule with the `TimestampLessThanEqualsPath` operator.

Usage

ChoiceRule$TimestampLessThanEqualsPath(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(str): Path to the value to compare `variable` against.

Returns

Rule: Rule with `TimestampLessThanEqualsPath` operator.

Method TimestampGreaterThanEquals()

Creates a rule with the `TimestampGreaterThanEquals` operator.

Usage

ChoiceRule$TimestampGreaterThanEquals(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(str): Constant value to compare `variable` against.

Returns

Rule: Rule with `TimestampGreaterThanEquals` operator.

Method TimestampGreaterThanEqualsPath()

Creates a rule with the `TimestampGreaterThanEqualsPath` operator.

Usage

ChoiceRule$TimestampGreaterThanEqualsPath(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(str): Path to the value to compare `variable` against.

Returns

Rule: Rule with `TimestampGreaterThanEqualsPath` operator.

Method IsNull()

Creates a rule with the `IsNull` operator.

Usage

ChoiceRule$IsNull(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(bool): Whether the value at `variable` is equal to the JSON literal null or not.

Returns

Rule: Rule with `IsNull` operator.

Method IsPresent()

Creates a rule with the `IsPresent` operator.

Usage

ChoiceRule$IsPresent(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(bool): Whether a field at `variable` exists in the input or not.

Returns

Rule: Rule with `IsPresent` operator.

Method IsString()

Creates a rule with the `IsString` operator.

Usage

ChoiceRule$IsString(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(bool): Whether the value at `variable` is a string or not.

Returns

Rule: Rule with `IsString` operator.

Method IsNumeric()

Creates a rule with the `IsNumeric` operator.

Usage

ChoiceRule$IsNumeric(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(bool): Whether the value at `variable` is a number or not.

Returns

Rule: Rule with `IsNumeric` operator.

Method IsTimestamp()

Creates a rule with the `IsTimestamp` operator.

Usage

ChoiceRule$IsTimestamp(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(bool): Whether the value at `variable` is a timestamp or not.

Returns

Rule: Rule with `IsTimestamp` operator.

Method IsBoolean()

Creates a rule with the `IsBoolean` operator.

Usage

ChoiceRule$IsBoolean(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(bool): Whether the value at `variable` is a boolean or not.

Returns

Rule: Rule with `IsBoolean` operator.

Method StringMatches()

Creates a rule with the `StringMatches` operator.

Usage

ChoiceRule$StringMatches(variable, value)

Arguments

variable

(str): Path to the variable to compare.

value

(str): A string pattern that may contain one or more `*` characters to compare the value at `variable` to. The `*` character can be escaped using two backslashes. The comparison yields true if the variable matches the pattern, where `*` is a wildcard that matches zero or more characters.

Returns

Rule: Rule with `StringMatches` operator.

Method And()

Creates a compound rule with the `And` operator.

Usage

ChoiceRule$And(rules)

Arguments

rules

(list(BaseRule)): List of rules to compound together.

Returns

CompoundRule: Compound rule with `And` operator.

Method Or()

Creates a compound rule with the `Or` operator.

Usage

ChoiceRule$Or(rules)

Arguments

rules

(list(BaseRule)): List of rules to compound together.

Returns

CompoundRule: Compound rule with `Or` operator.

Method Not()

Creates a negation for a rule.

Usage

ChoiceRule$Not(rule)

Arguments

rule

(BaseRule): Rule to Negate.

Returns

NotRule: Rule with `Not` operator.

Method clone()

The objects of this class are cloneable with this method.

Usage

ChoiceRule$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.