Map state provides the ability to dynamically iterate over a state/subgraph for each entry in a list.

A Map state can accept an input with a list of items, execute a state or chain for each item in the list, and return a list, with all corresponding results of each execution, as its output.

Super classes

stepfunctions::Block -> stepfunctions::State -> Map

Methods

Public methods

Map$new()

Map$attach_iterator()

Map$allowed_fields()

Map$to_list()

Map$clone()

Inherited methods

stepfunctions::Block$is_field_allowed()

stepfunctions::Block$to_json()

stepfunctions::State$.next()

stepfunctions::State$accept()

stepfunctions::State$add_catch()

stepfunctions::State$add_retry()

stepfunctions::State$format()

stepfunctions::State$output()

stepfunctions::State$update_parameters()

Method new()

Initialize Map state class

Usage

Map$new(
  state_id,
  iterator,
  items_path = NULL,
  max_concurrency = NULL,
  comment = NULL,
  input_path = NULL,
  parameters = NULL,
  result_path = NULL,
  output_path = NULL,
  ...
)

Arguments

state_id

(str): State name whose length **must be** less than or equal to 128 unicode characters. State names **must be** unique within the scope of the whole state machine.

iterator

(State or Chain): State or chain to execute for each of the items in `items_path`.

items_path

(str, optional): Path in the input for items to iterate over. (default: '$')

max_concurrency

(int, optional): Maximum number of iterations to have running at any given point in time. (default: 0)

comment

(str, optional): Human-readable comment or description. (default: None)

input_path

(str, optional): Path applied to the state’s raw input to select some or all of it; that selection is used by the state. (default: '$')

parameters

(list, optional): The value of this field becomes the effective input for the state.

result_path

(str, optional): Path specifying the raw input’s combination with or replacement by the state’s result. (default: '$')

output_path

(str, optional): Path applied to the state’s output after the application of `result_path`, producing the effective output which serves as the raw input for the next state. (default: '$')

...

: Extra field names to pass to Block class

Method attach_iterator()

Attach `State` or `Chain` as iterator to the Map state, that will execute for each of the items in `items_path`. If an iterator was attached previously with the Map state, it will be replaced.

Usage

Map$attach_iterator(iterator)

Arguments

iterator

(State or Chain): State or Chain to attach as iterator to the Map state.

Method allowed_fields()

allowed extra fields

Usage

Map$allowed_fields()

Method to_list()

Convert class to named list

Usage

Map$to_list()

Method clone()

The objects of this class are cloneable with this method.

Usage

Map$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.