WorkflowTemplate Class

Abstract class to create a template for Sagemaker workflows

Methods

Public methods

WorkflowTemplate$new()

WorkflowTemplate$render_graph()

WorkflowTemplate$get_workflow()

WorkflowTemplate$build_workflow_definition()

WorkflowTemplate$create()

WorkflowTemplate$execute()

WorkflowTemplate$format()

WorkflowTemplate$clone()

Method new()

Usage

WorkflowTemplate$new(s3_bucket, workflow, role, client, ...)

Arguments

s3_bucket

(str): S3 bucket under which the output artifacts from the training job will be stored. The parent path used is built using the format: ``s3://s3_bucket/pipeline_name/models/job_name/``. In this format, `pipeline_name` refers to the keyword argument provided for TrainingPipeline. If a `pipeline_name` argument was not provided, one is auto-generated by the pipeline as `training-pipeline-<timestamp>`. Also, in the format, `job_name` refers to the job name provided when calling the :meth:`TrainingPipeline.run()` method.

workflow

: Workflow for AWS State Machine.

role

(str): An AWS IAM role (either name or full Amazon Resource Name (ARN)). This role is used to create, manage, and execute the Step Functions workflows.

client

(SFN.Client, optional): sfn client to use for attaching the existing workflow in Step Functions to the Workflow object. If not provided, a default sfn client for Step Functions will be automatically created and used. (default: None)

...

: currently not implemented

Method render_graph()

Renders a visualization of the workflow graph.

Usage

WorkflowTemplate$render_graph(portrait = FALSE)

Arguments

portrait

(bool, optional): Boolean flag set to `True` if the workflow graph should be rendered in portrait orientation. Set to `False`, if the graph should be rendered in landscape orientation. (default: False)

Method get_workflow()

Returns Workflow

Usage

WorkflowTemplate$get_workflow()

Method build_workflow_definition()

Build the workflow definition for the inference pipeline with all the states involved.

Usage

WorkflowTemplate$build_workflow_definition()

Returns

:class:`~stepfunctions.steps.states.Chain`: Workflow definition as a chain of states involved in the the inference pipeline.

Method create()

Creates the workflow on Step Functions.

Usage

WorkflowTemplate$create()

Returns

str: The Amazon Resource Name (ARN) of the workflow created. If the workflow already existed, the ARN of the existing workflow is returned.

Method execute()

Run the inference pipeline.

Usage

WorkflowTemplate$execute(...)

Arguments

...

: Not yet implemented

Returns

:R:class:`~stepfunctions.workflow.Execution`: Running instance of the inference pipeline.

Method format()

format class

Usage

WorkflowTemplate$format()

Method clone()

The objects of this class are cloneable with this method.

Usage

WorkflowTemplate$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.